+++ ij2-common/src/main/java/imagej/data/Unsigned12BitType.java
@@ -96,7 +96,7 @@ public class Unsigned12BitType implements Type
 	}
 
 	@Override
-	public Object allocateStorageArray(int numPixels)
+	public Object allocateStorageArray(long numPixels)
 	{
 		long numStorageUnits = calcNumStorageUnitsFromPixelCount(numPixels);
